Add checking file exists in test

+++ chemicalchecker/tests/test_database.py
@@ -19,7 +19,8 @@ class TestDatabase(unittest.TestCase):
     def tearDown(self):
         test_dir = os.path.dirname(os.path.realpath(__file__))
         testfile = os.path.join(test_dir, 'test.sqlite')
-        os.remove(testfile)
+        if os.path.exists(testfile):
+            os.remove(testfile)
 
 
     def test_add_bulk(self):
